rename module to work-around maintaining previous re-exports
