+++ src/codecs/gif.rs
@@ -43,7 +43,7 @@ use crate::error::{
     UnsupportedError, UnsupportedErrorKind,
 };
 use crate::image::{AnimationDecoder, ImageDecoder, ImageFormat};
-use crate::io::Limits;
+use crate::Limits;
 use crate::traits::Pixel;
 use crate::ExtendedColorType;
 use crate::ImageBuffer;
@@ -96,7 +96,7 @@ impl<R: BufRead + Seek> ImageDecoder for GifDecoder<R> {
     }
 
     fn set_limits(&mut self, limits: Limits) -> ImageResult<()> {
-        limits.check_support(&crate::io::LimitSupport::default())?;
+        limits.check_support(&crate::LimitSupport::default())?;
 
         let (width, height) = self.dimensions();
         limits.check_dimensions(width, height)?;
