+++ src/image_reader/free_functions.rs
