+++ src/image_reader/image_reader.rs
